Updated possible meetings with future dates/Thursday requirement + read-me
DB To Do List:

A meeting with fewer than 3 mentees will be canceled on Friday. All participants of the meeting need to be notified of the cancelation. If a meeting has fewer than 2 mentors, the admin will need to add mentors to the meeting. For simplicity, you can just generate notification files which contain the names and email addresses of the users who should be notified.

Finish read-me that shows location of every requirement in PDF and how to perform them